+++ spec/features/group_features_spec.rb
@@ -167,6 +167,7 @@ RSpec.describe 'User interacts with Groups' do
   describe 'Group enrollments', js: true do
     before :each do
       @group = create :group
+      @other_group = create :group
       @unenrolled_students = create_list :student, 2, organization: @group.chapter.organization
     end
 
@@ -193,6 +194,49 @@ RSpec.describe 'User interacts with Groups' do
         expect(student.enrollments.count).to eql 1
         expect(student.enrollments.first.active_since.to_date.to_s).to eql 2.days.ago.to_date.to_s
       end
+    end
+
+    it 'gets message when there are no students which can be enrolled' do
+      visit "/groups/#{@other_group.id}"
+      click_link 'Enroll Students'
+
+      within('#modal') do
+        expect(page).to have_content 'No students to be enrolled'
+      end
+    end
+
+    it 'alerts if the group has a student with grades before their enrollment' do
+      group = create :group
+      student = create :student, organization: group.chapter.organization
+      enrollment = create :enrollment, student: student, group: group, active_since: 1.day.ago
+      lesson = create :lesson, group: group, date: 2.days.ago
+      create :grade, lesson: lesson, student: student
+      enrolled_since_date_formatted = enrollment.active_since.strftime('%Y-%m-%d')
+      lesson_date_formatted = lesson.date.strftime('%Y-%m-%d')
+      expected_alert_text = "Student graded before enrollment. Enrolled since \"#{enrolled_since_date_formatted}\" but has grades for \"#{lesson_date_formatted}\""
+
+      visit "/groups/#{group.id}"
+
+      expect(page).to have_content 'Students graded outside enrollment'
+      expect(page).to have_content 'Some students have grades outside their enrollment, please review them below'
+      expect(page).to have_selector('span.group > .tooltip', visible: :all, text: expected_alert_text)
+    end
+
+    it 'alerts if the group has a student with grades after their enrollment' do
+      group = create :group
+      student = create :student, organization: group.chapter.organization
+      enrollment = create :enrollment, student: student, group: group, active_since: 2.days.ago, inactive_since: 1.day.ago
+      lesson = create :lesson, group: group, date: Time.zone.today
+      create :grade, lesson: lesson, student: student
+      enrollment_ended_since_date_formatted = enrollment.inactive_since.strftime('%Y-%m-%d')
+      lesson_date_formatted = lesson.date.strftime('%Y-%m-%d')
+      expected_alert_text = "Student graded after enrollment. Enrollment ended since \"#{enrollment_ended_since_date_formatted}\" but has grades for \"#{lesson_date_formatted}\""
+
+      visit "/groups/#{group.id}"
+
+      expect(page).to have_content 'Students graded outside enrollment'
+      expect(page).to have_content 'Some students have grades outside their enrollment, please review them below'
+      expect(page).to have_selector('span.group > .tooltip', visible: :all, text: expected_alert_text)
     end
   end
 end